Refactor and tdmpc2 online corrections
