+++ stable_worldmodel/wm/tdmpc2/__init__.py
@@ -1,0 +1,1 @@
+from .tdmpc2 import *  # noqa: F403
